Cataract Eye disease detection using ML/requirement.txt
pillow
streamlit
pandas==2.0.3
numpy==1.25.2
scikit-learn==1.2.2
beautifulsoup4
tensorflow==2.15.0
requests